code/server/requirements.txt
asgiref==3.6.0
Django==4.2
django-dump-load-utf8==0.0.4
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.4.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.6
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tzdata==2022.7
urllib3==1.26.15
